+++ jwst/tweakreg/tweakreg_step.py
@@ -51,6 +51,7 @@ class TweakRegStep(Step):
         gaia_catalog = option('GAIADR2', 'GAIADR1', default='GAIADR2')
         min_gaia = integer(min=0, default=5) # Min number of GAIA sources needed
         save_gaia_catalog = boolean(default=False)  # Write out GAIA catalog as a separate product
+        output_use_model = boolean(default=True)  # When saving use `DataModel.meta.filename`
     """
 
     reference_file_types = []
@@ -70,6 +71,9 @@ class TweakRegStep(Step):
             # entries when aligning to GAIA
             self.expand_refcat = True
 
+        if len(images) == 0:
+            raise ValueError("Input must contain at least one image model.")
+
         # Build the catalogs for input images
         for image_model in images:
             catalog = make_tweakreg_catalog(
@@ -77,11 +81,10 @@ class TweakRegStep(Step):
                 brightest=self.brightest, peakmax=self.peakmax
             )
 
-            # filter out sources outside the image array if WCS validity
-            # region is provided:
-            wcs_bounds = image_model.meta.wcs.pixel_bounds
-            if wcs_bounds is not None:
-                ((xmin, xmax), (ymin, ymax)) = wcs_bounds
+            # filter out sources outside the WCS bounding box
+            bb = image_model.meta.wcs.bounding_box
+            if bb is not None:
+                ((xmin, xmax), (ymin, ymax)) = bb
                 xname = 'xcentroid' if 'xcentroid' in catalog.colnames else 'x'
                 yname = 'ycentroid' if 'ycentroid' in catalog.colnames else 'y'
                 x = catalog[xname]
@@ -116,11 +119,9 @@ class TweakRegStep(Step):
                               .format(catalog_filename))
                 image_model.meta.tweakreg_catalog = catalog_filename
 
+            # Temporarily attach catalog to the image model so that it follows
+            # the grouping by exposure, to be removed after use below
             image_model.catalog = catalog
-
-        # Now use the catalogs for tweakreg
-        if len(images) == 0:
-            raise ValueError("Input must contain at least one image model.")
 
         # group images by their "group id":
         grp_img = list(images.models_grouped)
@@ -143,6 +144,8 @@ class TweakRegStep(Step):
             self.skip = True
             for model in images:
                 model.meta.cal_step.tweakreg = "SKIPPED"
+                # Remove the attached catalogs
+                del model.catalog
             return input
 
         # create a list of WCS-Catalog-Images Info and/or their Groups:
@@ -153,6 +156,9 @@ class TweakRegStep(Step):
             else:
                 group_name = _common_name(g)
                 wcsimlist = list(map(self._imodel2wcsim, g))
+                # Remove the attached catalogs
+                for model in g:
+                    del model.catalog
                 self.log.info("* Images in GROUP '{}':".format(group_name))
                 for im in wcsimlist:
                     im.meta['group_id'] = group_name